Restructured cog modules
