+++ commanderbot/ext/help_chat/help_chat_state.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 from commanderbot.ext.help_chat.help_chat_guild_state import HelpChatGuildState
 from commanderbot.ext.help_chat.help_chat_store import HelpChatStore
-from commanderbot.lib import GuildPartitionedCogState
+from commanderbot.lib.cogs import GuildPartitionedCogState
 
 
 @dataclass
