+++ commanderbot/ext/help_forum/help_forum_json_store.py
@@ -5,7 +5,8 @@ from discord import ForumChannel, ForumTag, Guild
 
 from commanderbot.ext.help_forum.help_forum_data import HelpForumData
 from commanderbot.ext.help_forum.help_forum_store import HelpForum
-from commanderbot.lib import CogStore, JsonFileDatabaseAdapter
+from commanderbot.lib.cogs import CogStore
+from commanderbot.lib.cogs.database import JsonFileDatabaseAdapter
 
 
 # @implements HelpForumStore
