+++ commanderbot/ext/help_forum/help_forum_state.py
@@ -2,7 +2,7 @@ from dataclasses import dataclass
 
 from commanderbot.ext.help_forum.help_forum_guild_state import HelpForumGuildState
 from commanderbot.ext.help_forum.help_forum_store import HelpForumStore
-from commanderbot.lib import GuildPartitionedCogState
+from commanderbot.lib.cogs import GuildPartitionedCogState
 
 
 @dataclass
