+++ commanderbot/ext/invite/invite_options.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
 from typing import Any, Dict
 
-from commanderbot.lib import (
+from commanderbot.lib.cogs.database import (
     DatabaseOptions,
     InMemoryDatabaseOptions,
     make_database_options,
