--- commanderbot/lib/sql_database_adapter.py
@@ -1,16 +1,0 @@
-from typing import Protocol
-
-from sqlalchemy.ext.asyncio import AsyncConnection, AsyncSession
-
-__all__ = ("SQLDatabaseAdapter",)
-
-
-class SQLDatabaseAdapter(Protocol):
-    def connect(self) -> AsyncConnection:
-        ...
-
-    def begin(self) -> AsyncConnection:
-        ...
-
-    def session(self) -> AsyncSession:
-        ...
